AuthorizationService + MealService unit tests

+++ backend/Application/Services/MealService.cs
@@ -1,22 +1,13 @@
 ﻿using Application.Interfaces;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-
 using Domain;
 using WebApiContracts;
 using WebApiContracts.Mappers;
-using Microsoft.AspNetCore.Http;
 namespace Application.Services
 {
     public class MealService
     {
-
-        private IMealRepository _mealRepository;
-        private IFileRepository _fileRepository;
+        private readonly IMealRepository _mealRepository;
+        private readonly IFileRepository _fileRepository;
 
         public MealService(IMealRepository mealRepository, IFileRepository fileRepository) 
         {
@@ -26,35 +17,46 @@ namespace Application.Services
 
         public bool addMeal(Meal meal,Guid id)
         {
-           
             return _mealRepository.addMeal(meal, id);
         }
 
         public bool UpdateMeal(MealDTO meal)
         {
             return _mealRepository.UpdateMeal(meal.ToEntity());
-       
         }
 
         public MealDTO GetMealById(Guid id)
         {
+            var mealCheck = _mealRepository.GetMeal(id);
+
+            if (mealCheck == null)
+            {
+                throw new Exception("Meal doesn't exist!");
+            }
+
             var meal = _mealRepository.GetMeal(id);
             var file = _fileRepository.GetFile(id.ToString().ToUpper());
             var mealDTO = meal.ToDTO();
             mealDTO.Photo = file.FirstOrDefault().Path;
 
             return mealDTO;
-
         }
 
         public IEnumerable<MealDTO> GetAllMeals()
         {
             var meals  = _mealRepository.GetAllMeals();
             return meals.Select(x => x.ToDTO());
-
         }
+
         public bool DeleteMeal(Guid id)
         {
+            var mealCheck = _mealRepository.GetMeal(id);
+
+            if (mealCheck == null)
+            {
+                throw new Exception("Meal doesn't exist!");
+            }
+
             return _mealRepository.DeleteMeal(id);
         }
     }
